blood_test_analysis/requirements.txt
Flask
google-generativeai
python-dotenv
PyMuPDF
Pillow
pytesseract